Move providers to a separate sub-directory in all-devices-common
