+++ examples/all-devices-app/all-devices-common/providers/DeviceInfoProviderImpl.cpp
