+++ examples/all-devices-app/all-devices-common/providers/OverrideDeviceInstanceInfoProvider.cpp
