+++ examples/all-devices-app/posix/BUILD.gn
@@ -24,8 +24,8 @@ executable("all-devices-app") {
   ]
 
   deps = [
-    "${chip_root}/examples/all-devices-app/all-devices-common:device-info-provider",
-    "${chip_root}/examples/all-devices-app/all-devices-common:override-device-instance-info-provider",
+    "${chip_root}/examples/all-devices-app/all-devices-common/providers:device-info-provider",
+    "${chip_root}/examples/all-devices-app/all-devices-common/providers:override-device-instance-info-provider",
     "${chip_root}/examples/all-devices-app/all-devices-common/devices/boolean-state-sensor",
     "${chip_root}/examples/all-devices-app/all-devices-common/devices/chime",
     "${chip_root}/examples/all-devices-app/all-devices-common/devices/device-factory",
